ProbablyGenBackend/requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
Django==2.2.6
django-cors-headers==3.1.1
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.5.0
Markdown==3.2.1
murmurhash==1.0.2
numpy==1.18.1
plac==1.1.3
preshed==3.0.2
psycopg2==2.8.4
python-dotenv==0.10.3
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.2.post1
six==1.14.0
spacy==2.2.4
sqlparse==0.3.1
srsly==1.0.2
thinc==7.4.0
tqdm==4.43.0
uritemplate==3.0.0
urllib3==1.25.8
wasabi==0.6.0
zipp==3.1.0
